requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
dnspython==2.8.0
python-dotenv==1.2.2
fastapi==0.135.3
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
